Move IconTwoLineViewHolder.kt to mobile module
