+++ mobile/src/main/res/layout/recyclerview_item_icon_two_line.xml
